requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
psycopg2
pandas